tests/test_models.py
import pytest

from languages_plus.models import Language
from pldp.models.agency import Agency

@pytest.fixture
def language()

@pytest.fixture
def agency():
    agency = Agency.objects.create(name='Test Agency',
                                   email='test@testagency.com')
