tests/test_requirements.txt
-e .
flake8
pytest-django==3.4.4
psycopg2==2.7.6.1
